KD-Tree construction seems to work (but only if leaf size > 16 ?! WTF git diff

+++ raymini/KdTree.h
@@ -31,7 +31,7 @@ public:
 
 
 private:
-    uint8_t splitAxis;
+    unsigned int splitAxis;
 	float splitPosition;
 	BoundingBox bbox;
     KdTree *left;
